requirements.txt
asgiref==3.4.1
boto3==1.20.14
botocore==1.23.14
Deprecated==1.2.13
Django==3.2.9
django-model-utils==4.2.0
django-redis==5.0.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
jmespath==0.10.0
mysqlclient==2.1.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
wrapt==1.13.3
